requirements.txt
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
idna==2.8
instagram-scraper==1.5.41
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.1
opencv-python==4.0.0.21
pandas==0.24.1
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
torch==1.0.1.post2
tqdm==4.31.1
urllib3==1.24.1
